==> notebooks/kaggle_data/2_Genres.ipynb <==
import pandas as pd
import numpy as np

without_dub_union = pd.read_csv("without_dub_union.csv")

#### 9.1.Split the Genres columns

##### Genre1 to genre1_1 (G1, G2, G3)

genre_1 = without_dub_union[['Genre1']].copy()
genre_1[['G1', 'G2', 'G3']] = genre_1['Genre1'].str.split(',', expand=True)
genre1_1= genre_1.drop(['Genre1'], axis=1)
genre1_1

##### Genre 2 to genre2_1 (G4, G5, G6)

genre_2 = without_dub_union[['Genre2']].copy()
genre_2[['G4', 'G5', 'G6']] = genre_2['Genre2'].str.split(',', expand=True)
genre2_1= genre_2.drop(['Genre2'], axis=1)
genre2_1

##### Genre3 to genre3_1 ( G7, G8, G9)

genre_3 = without_dub_union[['Genre3']].copy()
genre_3[['G7', 'G8', 'G9']] = genre_3['Genre3'].str.split(',', expand=True)
genre3_1= genre_3.drop(['Genre3'], axis=1)
genre3_1

##### Genre4 to genre4_1 (G10, G11, G12)

genre_4 = without_dub_union[['Genre4']].copy()
genre_4[['G10', 'G11', 'G12']] = genre_4['Genre4'].str.split(',', expand=True)
genre4_1= genre_4.drop(['Genre4'], axis=1)
genre4_1

##### Genre5 to genre5_1 (G16)

without_dub_union.Genre5.unique()

without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("/",",")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("-"," ")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("Science fiction","Sci-Fi")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("Hidden camera","Hidden-camera")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace(" ",",")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("thriller","Thriller,")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("drama","Drama")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("comedy","Comedy")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("horror","Horror")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("adventure","Adventure")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("fantasy","Fantasy")
without_dub_union.Genre5 = without_dub_union.Genre5.str.replace("prank","Prank")

without_dub_union.Genre5

##### Genre6 to genre6_1 (G13, G14, G15)

genre_6 = without_dub_union[['Genre6']].copy()
genre_6[['G13', 'G14','G15']] = genre_6['Genre6'].str.split(',', expand=True)
genre6_1= genre_6.drop(['Genre6'], axis=1)
genre6_1

#### 9.2. Changes with unique(), concat and union

#### 9.2.1

##### genre1_1

genre1_1.columns

genre1_1.fillna('empty', inplace=True)

genre1_1['G1'] = genre1_1['G1'].astype("string")

genre1_1.G1.unique()

genre1_1.G2.unique()

genre1_1['G2'] = genre1_1['G2'].astype("string")

genre1_1.G2 = genre1_1['G2'].str.replace(' ', '')

genre1_1.G3.unique()

genre1_1.G3 = genre1_1['G3'].str.replace(' ', '')

genre1_1['G3'] = genre1_1['G3'].astype("string")
genre1_1.G3

##### genre2_1

genre2_1.columns
genre2_1.fillna('empty', inplace=True)

genre2_1.G4.unique()
genre2_1['G4'] = genre2_1['G4'].astype("string")

genre2_1.G5.unique()

genre2_1.G5 = genre2_1['G5'].str.replace(' ', '')
genre2_1['G5'] = genre2_1['G5'].astype("string")

genre2_1.G6.unique()

genre2_1.G6 = genre2_1['G6'].str.replace(' ', '')
genre2_1['G6'] = genre2_1['G6'].astype("string")

##### genres1_1 and genres2_1 concat (genres_concat1)

genres_concat1 = pd.concat([genre1_1.G1,genre1_1.G2,genre1_1.G3,genre2_1.G4,genre2_1.G5,genre2_1.G6], axis=1)
genres_concat1

##### genres_union1 

genres_concat1["genres_union1"] = (
    genres_concat1[['G1', 'G2', 'G3', 'G4', 'G5', 'G6']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
genres_concat1.genres_union1

#### 9.2.2

##### genre3_1

genre3_1.columns

genre3_1.fillna('empty', inplace=True)

genre3_1.G7.unique()

genre3_1['G7'] = genre3_1['G7'].astype("string")

genre3_1.G8.unique()

genre3_1['G8'] = genre3_1['G8'].astype("string")

genre3_1.G9.unique()

genre3_1['G9'] = genre3_1['G9'].astype("string")

##### genre4_1

genre4_1.columns

genre4_1.fillna('empty', inplace=True)

genre4_1.G10.unique()

genre4_1['G10'] = genre4_1['G10'].astype("string")

genre4_1.G11.unique()

genre4_1['G11'] = genre4_1['G11'].astype("string")

genre4_1.G11 = genre4_1['G11'].str.replace(' ', '')

genre4_1.G12.unique()

genre4_1.G12 = genre4_1['G12'].str.replace(' ', '')

genre4_1['G11'] = genre4_1['G11'].astype("string")

##### genres3_1 and genres4_1 concat (genres_concat2)

genres_concat2 = pd.concat([genre3_1.G7,genre3_1.G8,genre3_1.G9,genre4_1.G10,genre4_1.G11,genre4_1.G12], axis=1)
genres_concat2

##### genres_union2 

genres_concat2["genres_union2"] = (
    genres_concat2[['G7', 'G8', 'G9', 'G10', 'G11', 'G12']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
genres_concat2.genres_union2

#### 9.2.3

##### genre6_1

genre6_1.columns

genre6_1.fillna('empty', inplace=True)

genre6_1['G13'] = genre6_1['G13'].astype("string")

genre6_1.G13.unique()

genre6_1.G14.unique()

genre6_1['G14'] = genre6_1['G14'].astype("string")

genre6_1.G14 = genre6_1['G14'].str.replace(' ', '')

genre6_1.G15.unique()

genre6_1['G15'] = genre6_1['G15'].astype("string")

genre6_1.G15 = genre6_1['G15'].str.replace(' ', '')

##### Genre5

without_dub_union.Genre5.fillna('empty', inplace=True)

without_dub_union['Genre5'] = without_dub_union['Genre5'].astype("string")
without_dub_union.Genre5

#####  Genre5 and genres6_1 concat (genres_concat3)

genres_concat3 = pd.concat([genre6_1.G13,genre6_1.G14,genre6_1.G15,without_dub_union.Genre5], axis=1)
genres_concat3

##### genres_union3

genres_concat3["genres_union3"] = (
    genres_concat3[['G13', 'G14', 'G15', 'Genre5']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
genres_concat3.genres_union3

#### 9.2.4

##### genres_concat4

genres_concat4 = pd.concat([genres_concat1.genres_union1,genres_concat2.genres_union2, genres_concat3.genres_union3], axis=1)
genres_concat4

##### genres_union4

genres_concat4["genres_union4"] = (
    genres_concat4[['genres_union1', 'genres_union2', 'genres_union3']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
genres_concat4.genres_union4

#### 9.3 Split the genres_union4 and union with the without_dub_union.csv

##### Split

g_split = genres_concat4[['genres_union4']].copy()
g_split[['Genre_1', 'Genres_2', 'Genre_3', 'Genre_4', 'Genre_5',  'Genre_6' ]] = g_split['genres_union4'].str.split(',', expand=True)
g_split= g_split.drop(['genres_union4'], axis=1)
g_split

g_split.fillna('', inplace=True)
g_split

##### union with .csv

without_dub_union.columns

without_dub_union = without_dub_union.drop(['Genre1', 'Genre2', 'Genre3', 'Genre4', 'Genre5', 'Genre6' ], axis=1)
without_dub_union.columns

without_dub_union = without_dub_union.drop(['Unnamed: 0' ], axis=1)


classification_df = pd.concat([without_dub_union, g_split], axis=1)
classification_df.columns

classification_df.to_csv('movies_changes1.csv')
