==> notebooks/kaggle_data/3_Actors.ipynb <==
import pandas as pd
import numpy as np

movies_changes1 = pd.read_csv("movies_changes1.csv")

movies_changes1.columns

movies_changes1 = pd.read_csv("movies_changes1.csv")

#### 10. Union the columns of the movies_changes1.csv
#### (Actors, Actor1, Actor2, Actor3, Actor4, Star1, Star2, Star3, Star4)

##### 10.1 Split the Actors column

Actors_ = movies_changes1[['Actors']].copy()
Actors_[['Actor5', 'Actor6', 'Actor7', 'Actor8']] = Actors_['Actors'].str.split(',', expand=True)
Actors_= Actors_.drop(['Actors'], axis=1)
Actors_

Actors_.fillna('empty', inplace=True)

Actors_['Actor5'] = Actors_['Actor5'].astype("string")

Actors_.Actor5.unique()

Actors_['Actor6'] = Actors_['Actor6'].astype("string")

Actors_.Actor6.unique()

Actors_.Actor6.unique()

Actors_['Actor7'] = Actors_['Actor7'].astype("string")

Actors_.Actor7.unique()

Actors_['Actor8'] = Actors_['Actor8'].astype("string")

Actors_.Actor8.unique()

##### 10.2 checking the rest columns with unique()

movies_changes1.Actor1.unique()
movies_changes1['Actor1'] = movies_changes1['Actor1'].astype("string")

movies_changes1.Actor2.unique()
movies_changes1['Actor2'] = movies_changes1['Actor2'].astype("string")

movies_changes1.Actor3.unique()
movies_changes1['Actor3'] = movies_changes1['Actor3'].astype("string")

movies_changes1.Actor4.unique()
movies_changes1['Actor4'] = movies_changes1['Actor4'].astype("string")

movies_changes1.Star1.unique()
movies_changes1['Star1'] = movies_changes1['Star1'].astype("string")

movies_changes1.Star2.unique()
movies_changes1['Star2'] = movies_changes1['Star2'].astype("string")

movies_changes1.Star3.unique()
movies_changes1['Star3'] = movies_changes1['Star3'].astype("string")

movies_changes1.Star4.unique()
movies_changes1['Star4'] = movies_changes1['Star4'].astype("string")

##### 10.3 concat & union

##### actors_concat1 & actors_union1 

actors_concat1 = pd.concat([Actors_.Actor5,Actors_.Actor6,Actors_.Actor7,Actors_.Actor8], axis=1)
actors_concat1

actors_concat1["actors_union1"] = (
    actors_concat1[['Actor5', 'Actor6', 'Actor7', 'Actor8']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
actors_concat1.actors_union1

##### actors_concat2 & actors_union2

actors_concat2 = pd.concat([movies_changes1.Actor1,movies_changes1.Actor2,movies_changes1.Actor3,movies_changes1.Actor4], axis=1)
actors_concat2

actors_concat2.fillna('empty', inplace=True)
actors_concat2

actors_concat2["actors_union2"] = (
    actors_concat2[['Actor1', 'Actor2', 'Actor3', 'Actor4']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
actors_concat2.actors_union2

##### actors_concat3 & actors_union3

actors_concat3 = pd.concat([movies_changes1.Star1,movies_changes1.Star2,movies_changes1.Star3,movies_changes1.Star4], axis=1)
actors_concat3

actors_concat3.fillna('empty', inplace=True)
actors_concat3

actors_concat3["actors_union3"] = (
    actors_concat3[['Star1', 'Star2', 'Star3', 'Star4']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
actors_concat3.actors_union3

##### actors_concat4 & actors_union4 

actors_concat4 = pd.concat([actors_concat1.actors_union1,actors_concat2.actors_union2, actors_concat3.actors_union3], axis=1)
actors_concat4

actors_concat4["actors_union4"] = (
    actors_concat4[['actors_union1', 'actors_union2', 'actors_union3']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
actors_concat4.actors_union4

#### 10.4 split actors_concat1.actors_union1 

a_split = actors_concat4[['actors_union4']].copy()
a_split[['Actor_1', 'Actor_2', 'Actor_3', 'Actor_4']] = a_split['actors_union4'].str.split(',', expand=True)
a_split= a_split.drop(['actors_union4'], axis=1)
a_split

a_split.Actor_1 = a_split['Actor_1'].str.replace('empty', '')
a_split.fillna('', inplace=True)
a_split

#### 10.5 removing the old columns and replcae with the a_split to movies_change1 (movies_change2)

movies_changes1 = movies_changes1.drop(['Actor1', 'Actor2', 'Actor3', 'Actor4', 'Star1', 
                                        'Star2','Star3','Star4','Actors' ], axis=1)
movies_changes1.columns

movies_changes2 = pd.concat([movies_changes1, a_split], axis=1)
movies_changes2.columns

movies_changes2.to_csv('movies_changes2.csv')

movies_changes2.dtypes

print(movies_changes2)
