==> notebooks/kaggle_data/4_Certificates.ipynb <==
import pandas as pd
import numpy as np

movies_changes2 = pd.read_csv("movies_changes2.csv")

movies_changes2.columns

#### 11. Union the columns of the movies_changes2.csv (Certificate4, Certificate7, Certificate)

#### 11.1 cheking columns with unique() 

movies_changes2.Certificate4.unique()

movies_changes2['Certificate4'] = movies_changes2['Certificate4'].astype("string")

movies_changes2.Certificate7.unique()

movies_changes2['Certificate7'] = movies_changes2['Certificate7'].astype("string")

movies_changes2.Certificate.unique()

movies_changes2['Certificate'] = movies_changes2['Certificate'].astype("string")

#### 11.2 concat & union

##### 11.2.1 certificate_concat1

certificate_concat1 = pd.concat([movies_changes2.Certificate4,movies_changes2.Certificate7,movies_changes2.Certificate], axis=1)
certificate_concat1

certificate_concat1.fillna('empty', inplace=True)

##### 11.2.2 certificate_union1

certificate_concat1["certificate_union1"] = (
    certificate_concat1[['Certificate4', 'Certificate7', 'Certificate']]
        .apply(lambda x: x.str.split(',\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'empty'])))        
        .replace('', 'empty')
    )
certificate_concat1.certificate_union1

#### 11.3 split certificate_concat1.certificate_union1 

c_split = certificate_concat1[['certificate_union1']].copy()
c_split[['Certificate_1', 'Certificate_2']] = c_split['certificate_union1'].str.split(',', expand=True)
c_split= c_split.drop(['certificate_union1'], axis=1)
c_split

c_split.fillna('', inplace=True)
c_split

#### 11.4 removing the old columns and replcae with the c_split to movies_change2 (movies_change3)

movies_changes2 = movies_changes2.drop(['Certificate', 'Certificate4', 'Certificate7'], axis=1)
movies_changes2.columns

movies_changes3 = pd.concat([movies_changes2, c_split], axis=1)
movies_changes3.columns

movies_changes3= movies_changes3.drop(['Unnamed: 0.1', 'Unnamed: 0' ], axis=1)
movies_changes3.columns

movies_changes3.to_csv('movies_changes3.csv')

print(movies_changes3)
