==> notebooks/kaggle_data/5_Description.ipynb <==
import pandas as pd
import numpy as np

movies_changes3 = pd.read_csv("movies_changes3.csv")

movies_changes3.columns

#### 12. Union the columns of the movies_changes2.csv (Description1, Description2, Description3,Description4, Description5 )

#### 12.1 columns to string 

movies_changes3['Description1'] = movies_changes3['Description1'].astype("string")
movies_changes3.Description1.dtypes

movies_changes3['Description2'] = movies_changes3['Description2'].astype("string")
movies_changes3.Description2.dtypes

movies_changes3['Description3'] = movies_changes3['Description3'].astype("string")
movies_changes3.Description3.dtypes

movies_changes3['Description4'] = movies_changes3['Description4'].astype("string")
movies_changes3.Description4.dtypes

movies_changes3['Description5'] = movies_changes3['Description5'].astype("string")
movies_changes3.Description5.dtypes

#### 12.2 concat & union

##### 12.2.1 description_concat1

description_concat1 = pd.concat([movies_changes3.Description1,movies_changes3.Description2,
                                 movies_changes3.Description3, movies_changes3.Description4,
                                 movies_changes3.Description5], axis=1)
description_concat1

description_concat1.fillna('kkempty', inplace=True)
description_concat1

##### 12.2.2 certificate_union1

description_concat1["description_union1"] = (
    description_concat1[['Description1', 'Description2', 'Description3', 'Description4', 'Description5']]
        .apply(lambda x: x.str.split('<\s*'))
        .sum(axis=1).map(np.array)
        .apply(lambda x: ', '.join(np.unique(x[x != 'kkempty'])))        
        .replace('', 'kkempty')
    )
description_concat1.description_union1

#### 12.3 split description_concat1.description_union1

d_split = description_concat1[['description_union1']].copy()
d_split[['Description']] = d_split['description_union1'].str.split('<', expand=True)
d_split= d_split.drop(['description_union1'], axis=1)
d_split

#### 12.4 removing the old columns and replcae with the d_split to movies_change3 (movies_change4)

movies_changes3 = movies_changes3.drop(['Description1', 'Description2', 'Description3',
                                       'Description4', 'Description5'], axis=1)
movies_changes3.columns

movies_changes4 = pd.concat([movies_changes3, d_split], axis=1)
movies_changes4.columns

movies_changes4= movies_changes4.drop(['Unnamed: 0'], axis=1)
movies_changes4.columns

movies_changes4.Description = movies_changes4['Description'].str.replace('kkempty', '')

movies_changes4.to_csv('movies_changes4.csv')

print(movies_changes4)
