==> notebooks/kaggle_data/6_last_changes.ipynb <==
import pandas as pd
import numpy as np

movies_changes4= pd.read_csv("movies_changes4.csv")

movies_changes4.columns

movies_changes4.dtypes

movies_changes4.columns          

cols_to_str = ['Certificate_1', 'Certificate_2', 'Description',
               'Actor_1', 'Actor_2', 'Actor_3', 'Actor_4',
               'Genre_1', 'Genres_2', 'Genre_3', 'Genre_4', 'Genre_5', 'Genre_6',
               'Nudity', 'Violence', 'Profanity', 'Alcohol', 'Frightening',
               'Name', 'Type', 'Director']

movies_changes4[cols_to_str] = movies_changes4[cols_to_str].astype("string")

movies_changes4.Episodes.unique()

movies_changes4.Episodes = movies_changes4['Episodes'].str.replace('-', '')
movies_changes4.Episodes = pd.to_numeric(movies_changes4.Episodes)
movies_changes4.Episodes.dtypes

movies_changes4.Language.unique()


movies_changes4.Language = movies_changes4['Language'].str.replace('cn', 'Chinese')
movies_changes4.Language = movies_changes4['Language'].str.replace('/', ',')
movies_changes4['Language'] = movies_changes4['Language'].astype("string")

movies_changes4[['Language1', 'Language2']] = movies_changes4['Language'].str.split(',', expand=True)
movies_changes4= movies_changes4.drop(['Language'], axis=1)
movies_changes4.columns

movies_changes4= movies_changes4.drop(['Average_Votes'], axis=1)


movies_changes4 = movies_changes4.dropna(subset=['Description'])

movies_changes4 = movies_changes4.drop_duplicates(subset=['Name', 'Type', 'Description'], inplace=False)

movies_changes4.to_csv('movies_changes5.csv', index = False)

movies_changes5.Description.isnull().values.any()

print(movies_changes4)
